Maven-ize the project
Create basic pom.xml to allow compiling and building the project
withouht having to mess with the depenencies and easying the task of
opening the project in an IDE-indepdendent way (and CLI, of course).

Remove lib directory, which is no longer needed. hadoop-nfs lib is
replaced by hadoop-core dependency, and testng not added to pom.xml
since it seems is not currently used.

Copied dists.dss to the appropriate location in Maven so that it is
built into the package.

Added plugin to maven to generate a "fat jar", so that program can
easily be run with java -jar.